refactor: 공통 utils함수로 추출
