+++ frontend/techpick/src/utils/dnd/isDnDCurrentData.ts
