+++ frontend/techpick/src/utils/dnd/isSelectionActive.ts
@@ -1,0 +1,1 @@
+export const isSelectionActive = (length: number) => 0 < length;
