+++ frontend/techpick/src/utils/index.ts
@@ -4,3 +4,6 @@ export { getClientCookie } from './getClientCookie';
 export { hasIndex } from './array';
 export { isEmptyString } from './string';
 export { getPortalContainer } from './portal';
+export { isDnDCurrentData } from './dnd/isDnDCurrentData';
+export { reorderSortableIdList } from './dnd/reorderSortableList';
+export { isSelectionActive } from './dnd/isSelectionActive';
